Refactor the Allocator class.
